requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-jazzmin==3.0.0
django-import-export==3.3.6
openpyxl==3.1.2
tzdata==2023.4